rename and add some utils
