+++ api/testing/tester_test.go
@@ -1,4 +1,4 @@
-package tester
+package testing
 
 import (
 	"lmm/api/db"
@@ -8,21 +8,21 @@ import (
 )
 
 func TestInit(t *testing.T) {
-	tester := New(t)
+	tester := NewTester(t)
 	db := db.Default()
 	_, err := db.Query("SHOW TABLES")
 	tester.NoError(err)
 }
 
 func TestIs(t *testing.T) {
-	tester := New(t)
+	tester := NewTester(t)
 	tester.Is(1, 1)
 	tester.Is(true, true)
 	tester.Is("abc", "abc")
 }
 
 func TestNot(t *testing.T) {
-	tester := New(t)
+	tester := NewTester(t)
 	tester.Not(1, 2)
 	tester.Not(true, false)
 	tester.Not(1, "1")
@@ -30,12 +30,12 @@ func TestNot(t *testing.T) {
 }
 
 func TestError(t *testing.T) {
-	tester := New(t)
+	tester := NewTester(t)
 	tester.Error(errors.New(""))
 	tester.Error(errors.New("msg"))
 }
 
 func TestNoError(t *testing.T) {
-	tester := New(t)
+	tester := NewTester(t)
 	tester.NoError(nil)
 }
